endpoint validation
added regex that ensures endpoint set in config is a valid http endpoint

fix: moved mount endpoint validation to util/ip
