+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "tau-tower"
-version = "0.2.101"
+version = "0.2.2"
 edition = "2024"
 
 [dependencies]
@@ -20,6 +20,7 @@ clap = { version = "4.5.47", features = ["derive"] }
 is-ip = "0.1.0"
 openssl = { version = "0.10.73", features = ["vendored"] }
 inline_colorization = "0.1.6"
+regex-lite = "0.1.9"
 
 [target.x86_64-unknown-linux-gnu]
 linker = "x86_64-linux-gnu-gcc"
